Refactor CourseBlock and SearchResultBlock to use getCourseColor utility for color assignment
