+++ src/app/course-plan/[id]/components/CourseBlock.tsx
@@ -1,17 +1,9 @@
-import { useRef } from "react";
 import { CourseBasicInfo } from "../types/Course";
-import { useDrag } from "react-dnd";
 import clsx from "clsx";
 import { useState } from "react";
 import { useEffect } from "react";
 import DraggableBlock from "@/app/components/DraggableBlock";
-
-const levelColors: { [key: number]: string } = {
-  1: "bg-green-800/30 text-gray-80",
-  2: "bg-zinc-500/40",
-  3: "bg-teal-700/30",
-  4: "bg-stone-600/40 text-stone-800",
-};
+import { getCourseColor } from "../utils";
 
 interface CourseBlockProps {
   semesterPlanId: string;
@@ -22,28 +14,11 @@ export default function CourseBlock({
   semesterPlanId,
   course,
 }: CourseBlockProps) {
-  const drag = useRef<HTMLDivElement>(null);
-  const [{ isDragging }, dragConnector] = useDrag(() => ({
-    type: "COURSE",
-    item: { semesterPlanId, course },
-    collect: (monitor) => ({
-      isDragging: !!monitor.isDragging(),
-    }),
-  }));
-  dragConnector(drag);
-
   //set the color of course block base on number of semester
   const [color, setColor] = useState<string>("bg-neutral-200");
 
   useEffect(() => {
-    // Extract the first digit from the numeric part of the course code (e.g., CSCI2100 -> 2)
-    const match = course.code.match(/\d/);
-    const level = match ? parseInt(match[0]) : undefined;
-    setColor(
-      level !== undefined
-        ? levelColors[level] || "bg-neutral-200"
-        : "bg-neutral-200",
-    );
+    setColor(getCourseColor(course.code));
   }, [course.code]);
 
   return (
